do_sqlite_checkpoint function moved to utils (relates #93)
